#593: Move json authoring tools to text-editor component
