+++ CourageScores/ClientApp/src/components/admin/Templates.test.js
@@ -476,122 +476,5 @@ describe('Templates', () => {
 
             expect(context.container.textContent).not.toContain('Could not save template');
         });
-
-        it('does not transform single line excel whitespace input', async () => {
-            const template = {
-                id: createTemporaryId(),
-                name: 'TEMPLATE',
-                sharedAddresses: [],
-                divisions: [],
-            };
-            templates = [template];
-            await renderComponent();
-            await doClick(findButton(context.container, 'Add'));
-            await doClick(context.container, 'input[name="editorFormat"]'); // switch to text editor
-            const input = '   ';
-
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', input, context.user);
-
-            const textareaOutput = context.container.querySelector('textarea[placeholder="Copy into template"]');
-            expect(textareaOutput.value).toEqual('');
-        });
-
-        it('transforms single line excel fixture input correctly', async () => {
-            const template = {
-                id: createTemporaryId(),
-                name: 'TEMPLATE',
-                sharedAddresses: [],
-                divisions: [],
-            };
-            templates = [template];
-            await renderComponent();
-            await doClick(findButton(context.container, 'Add'));
-            await doClick(context.container, 'input[name="editorFormat"]'); // switch to text editor
-            const input = 'A\tB\t\tC\t\D';
-
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', input, context.user);
-
-            const textareaOutput = context.container.querySelector('textarea[placeholder="Copy into template"]');
-            expect(JSON.parse(textareaOutput.value)).toEqual({
-                fixtures: [
-                    {home: 'A', away: 'B'},
-                    {home: 'C', away: 'D'}
-                ]
-            });
-        });
-
-        it('transforms multi line excel fixture input correctly', async () => {
-            const template = {
-                id: createTemporaryId(),
-                name: 'TEMPLATE',
-                sharedAddresses: [],
-                divisions: [],
-            };
-            templates = [template];
-            await renderComponent();
-            await doClick(findButton(context.container, 'Add'));
-            await doClick(context.container, 'input[name="editorFormat"]'); // switch to text editor
-            const input = 'A\tB\t\tC\t\D\n' +
-                'E\tF\t\tG\tH\n\n';
-
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', input, context.user);
-
-            const textareaOutput = context.container.querySelector('textarea[placeholder="Copy into template"]');
-            expect(JSON.parse('[' + textareaOutput.value + ']')).toEqual([{
-                fixtures: [
-                    {home: 'A', away: 'B'},
-                    {home: 'C', away: 'D'}
-                ]
-            }, {
-                fixtures: [
-                    {home: 'E', away: 'F'},
-                    {home: 'G', away: 'H'}
-                ]
-            }]);
-        });
-
-        it('transforms single line excel bye fixture input correctly', async () => {
-            const template = {
-                id: createTemporaryId(),
-                name: 'TEMPLATE',
-                sharedAddresses: [],
-                divisions: [],
-            };
-            templates = [template];
-            await renderComponent();
-            await doClick(findButton(context.container, 'Add'));
-            await doClick(context.container, 'input[name="editorFormat"]'); // switch to text editor
-            const input = 'A\t-\t\tC\t\D';
-
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', input, context.user);
-
-            const textareaOutput = context.container.querySelector('textarea[placeholder="Copy into template"]');
-            expect(JSON.parse(textareaOutput.value)).toEqual({
-                fixtures: [
-                    {home: 'A'},
-                    {home: 'C', away: 'D'}
-                ]
-            });
-        });
-
-        it('clears transformed output when input is cleared', async () => {
-            const template = {
-                id: createTemporaryId(),
-                name: 'TEMPLATE',
-                sharedAddresses: [],
-                divisions: [],
-            };
-            templates = [template];
-            await renderComponent();
-            await doClick(findButton(context.container, 'Add'));
-            await doClick(context.container, 'input[name="editorFormat"]'); // switch to text editor
-            const input = 'A\t-\t\tC\t\D';
-
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', input, context.user);
-            await doChange(context.container, 'textarea[placeholder="Copy from excel"]', '', context.user);
-
-            const textareaOutput = context.container.querySelector('textarea[placeholder="Copy into template"]');
-            expect(textareaOutput.value).toEqual('');
-        });
     });
 });
